WNB-154: Extract signUp and signIn into separate reducer slices
