+++ src/sagas/auth/signin/email-signin.js
@@ -1,6 +1,6 @@
 import { call, put, takeEvery, takeLatest } from '@redux-saga/core/effects';
 
-import { AUTH_METHODS } from 'consts';
+import { AUTH_METHODS, ACCESS_TOKEN, AUTH_METHOD } from 'consts';
 import {
   loadUserFailure,
   loadUserSuccess,
@@ -9,37 +9,6 @@ import {
   USER_LOGIN_REQUEST,
   startLoginPasswordAuth,
 } from 'store';
-import { ACCESS_TOKEN, AUTH_METHOD } from '../../../constants';
-
-export function* loginPasswordLoginSaga(client) {
-  yield takeLatest(startLoginPasswordAuth, function* handler() {
-    try {
-      const { user, token, error } = yield call(
-        [client, 'request'],
-        // TODO: do not hardcode url here, get it from env
-        'http://localhost:5000/users/login',
-        {
-          method: 'POST',
-        }
-      );
-
-      if (error) {
-        console.error(error);
-        // TODO handle it on ui somehow
-      }
-
-      yield put(loadUserSuccess(user));
-
-      if (token) {
-        localStorage.setItem('accessToken', token);
-        localStorage.setItem('authMethod', AUTH_METHODS.LOGIN_PASSWORD);
-      }
-    } catch (error) {
-      console.error(error);
-      yield put(loadUserFailure(error));
-    }
-  });
-}
 
 export function* emailLoginSaga(client) {
   yield takeEvery(USER_LOGIN_REQUEST, logIn, client);
