+++ src/sagas/auth/signup/email.js
@@ -1,46 +1,11 @@
-import { call, takeLatest, put, takeEvery } from '@redux-saga/core/effects';
+import { call, put, takeEvery } from '@redux-saga/core/effects';
 
 import { AUTH_METHODS, ACCESS_TOKEN, AUTH_METHOD } from 'consts';
 import {
-  startLoginPasswordSignup,
-  loadUserFailure,
-  loadUserSuccess,
   signUpRequestFailure,
   signUpRequestSuccess,
   USER_SIGN_UP_REQUEST,
 } from 'store';
-
-export function* loginPasswordSignUpSaga(client) {
-  yield takeLatest(startLoginPasswordSignup, function* handler() {
-    try {
-      const { user, token, error } = yield call(
-        [client, 'post'],
-        // TODO: do not hardcode url here, get it from env
-        'http://localhost:5000/users/signup',
-        {
-          data: {
-            login,
-            password,
-          },
-        }
-      );
-
-      if (error) {
-        console.error(error);
-      }
-
-      yield put(loadUserSuccess(user));
-
-      if (token) {
-        localStorage.setItem('accessToken', token);
-        localStorage.setItem('authMethod', AUTH_METHODS.LOGIN_PASSWORD);
-      }
-    } catch (error) {
-      console.error(error);
-      yield put(loadUserFailure(error));
-    }
-  });
-}
 
 export function* signUpSaga(client) {
   yield takeEvery(USER_SIGN_UP_REQUEST, signUp, client);
