+++ src/store/auth/index.js
@@ -4,3 +4,5 @@ export * from './forgotPassword';
 export * from './updateUserInfo';
 export * from './user';
 export * from './actions';
+export * from './signIn';
+export * from './signUp';
