+++ src/store/auth/signIn/constants.js
@@ -1,0 +1,11 @@
+export const USER_LOGIN_REQUEST = 'USER_LOGIN_REQUEST';
+export const USER_LOGIN_REQUEST_SUCCESS = 'USER_LOGIN_REQUEST_SUCCESS';
+export const USER_LOGIN_REQUEST_FAILURE = 'USER_LOGIN_REQUEST_FAILURE';
+
+export const CLEAR_LOGIN_ERROR = 'CLEAR_LOGIN_ERROR';
+
+export const INITIAL_STATE = {
+  loading: false,
+  loaded: false,
+  error: '',
+};
