+++ src/store/auth/signIn/selectors.js
@@ -1,0 +1,4 @@
+import * as R from 'ramda';
+
+export const getSignInError = state =>
+  R.pathOr(null, ['auth', 'signIn', 'error'], state);
