+++ src/store/auth/signUp/constants.js
@@ -1,0 +1,11 @@
+export const USER_SIGN_UP_REQUEST = 'USER_SIGN_UP_REQUEST';
+export const USER_SIGN_UP_REQUEST_SUCCESS = 'USER_SIGN_UP_REQUEST_SUCCESS';
+export const USER_SIGN_UP_REQUEST_FAILURE = 'USER_SIGN_UP_REQUEST_FAILURE';
+
+export const CLEAR_SIGNUP_ERROR = 'CLEAR_SIGNUP_ERROR';
+
+export const INITIAL_STATE = {
+  loading: false,
+  loaded: false,
+  error: '',
+};
