+++ src/store/auth/signUp/selectors.js
@@ -1,0 +1,4 @@
+import * as R from 'ramda';
+
+export const getSignUpError = state =>
+  R.pathOr(null, ['auth', 'signUp', 'error'], state);
