+++ src/store/auth/user/index.js
@@ -1,3 +1,4 @@
 export * from './actionCreators';
 export * from './constants';
 export * from './selectors';
+export * from './reducer';
