+++ src/ui-kit/theme/default/variables.js
@@ -6,7 +6,7 @@ export const mainBackgroundColor = '#F5F6FA';
 
 export const mainFontColor = '#000000';
 
-export const fontSmall = '10px';
+export const fontSmall = '12px';
 
 //colors
 export const lightGrey = '#E0E0E0';
